requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.9
# by the following command:
#
#    pip-compile
#
appdirs==1.4.4
    # via requests-cache
asgiref==3.6.0
    # via django
attrs==22.2.0
    # via
    #   cattrs
    #   pytest
    #   requests-cache
cattrs==22.2.0
    # via requests-cache
certifi==2022.12.7
    # via requests
charset-normalizer==3.0.1
    # via requests
coverage==7.1.0
    # via -r requirements.in
django==4.1.7
    # via -r requirements.in
exceptiongroup==1.1.0
    # via
    #   cattrs
    #   pytest
idna==3.4
    # via requests
iniconfig==2.0.0
    # via pytest
packaging==23.0
    # via pytest
petl==1.7.12
    # via -r requirements.in
pluggy==1.0.0
    # via pytest
psycopg2==2.9.5
    # via -r requirements.in
pympler==1.0.1
    # via -r requirements.in
pytest==7.2.1
    # via -r requirements.in
python-dateutil==2.8.2
    # via -r requirements.in
requests==2.28.2
    # via
    #   -r requirements.in
    #   requests-cache
requests-cache==0.9.8
    # via -r requirements.in
six==1.16.0
    # via
    #   python-dateutil
    #   url-normalize
sqlparse==0.4.3
    # via django
tomli==2.0.1
    # via pytest
url-normalize==1.4.3
    # via requests-cache
urllib3==1.26.14
    # via
    #   requests
    #   requests-cache
